Adds organization segment to module generation
Adds the ability to specify an organization segment in the
module generation dialog. This allows users to customize the
package name of the generated modules.

The organization segment is added to the module scaffolder, which
now uses it to create the package name for the generated modules.
The Gradle build file template is also updated to use the new
package name.